backend/requirements.txt
asgiref==3.5.1
Django==4.0.4
django-cors-headers==3.12.0
django-filter==22.1
djangorestframework==3.13.1
mysqlclient==2.1.1
python-decouple==3.6
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
python-decouple==3.6

